requirements.txt
asgiref==3.4.1
Django==3.2.8
django-braces==1.15.0
django-ckeditor==6.1.0
django-comments-dab==2.7.1
django-js-asset==1.2.2
django-model-utils==4.2.0
django-star-ratings==0.9.1
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
swapper==1.1.2.post1
Unidecode==1.3.2
